introduce new select dsl
existing functionality moved to alts function
